requirements.txt
# Core dependencies
apache-airflow[api]==2.8.0
apache-airflow-providers-apache-spark==4.5.0
pyspark==3.5.0
requests==2.31.0
pandas==2.0.3

# Storage and data format
boto3==1.33.13
s3fs==2023.12.2
pyarrow==11.0.0

# Data validation and schema
pydantic==2.5.2
marshmallow==3.20.1
great-expectations==0.18.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-asyncio==0.23.2
requests-mock==1.11.0
moto==4.2.11

# Code quality
black==24.1a1
isort==5.12.0
flake8==6.1.0
mypy==1.2.0
pre-commit==3.5.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development tools
ipython==8.12.3
python-dotenv==1.0.0

